pyproject.toml
[project]
name = "hello"
version = "0.1.0"
description = "An example Django app running in Docker."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "django==5.2.1",
  "celery==5.5.3",
  "django-debug-toolbar==5.2.0",
  "gunicorn==23.0.0",
  "psycopg==3.2.9",
  "redis==6.2.0",
  "ruff==0.11.12",
  "setuptools==80.9.0",
  "whitenoise==6.9.0",
  "djangorestframework==3.15.1"
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
extend-select = ["I", "SIM"]
